Refactored move logic. Only one robot moves, skeleton for attack and reproduce is there. Next, add many obstacles.

+++ PlayerRobot.h
@@ -8,7 +8,8 @@ class PlayerRobot: public Player
 {
    public:
       PlayerRobot();
-      PlayerRobot(int x, int y, Player_Type type, int powerLevel, int feed, int mate, int life);
+      PlayerRobot(short int x, short int y, Player_Type type, short int powerLevel,
+                  short int feed, short int mate, short int life);
 
       int getPower();
       int getTimeToFeed();
@@ -21,10 +22,10 @@ class PlayerRobot: public Player
       virtual ~PlayerRobot();
 
    private:
-      int power;
-      int timeToFeed;
-      int timeToReproduce;
-      int timeLifeRemaining;
+      short int power;
+      short int timeToFeed;
+      short int timeToReproduce;
+      short int timeLifeRemaining;
 
       // In increasing numerical order of priority
       typedef enum
